restructure encryption
