+++ .github/workflows/test.yml
@@ -101,24 +101,6 @@ jobs:
       - name: Run rust tests
         run: cargo test
 
-  rust-coverage-of-doc:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: RDXWorks-actions/checkout@main
-
-      - uses: RDXWorks-actions/toolchain@master
-        with:
-          toolchain: ${{ env.RUST_TOOLCHAIN }}
-          target: x86_64-unknown-linux-gnu
-          default: true
-
-      - name: manual install target
-        run: rustup target add ${{ env.RUST_TOOLCHAIN }}-x86_64-unknown-linux-gnu
-
-      - id: doccoverage
-        uses: RDXWorks-actions/rustdoc-coverage-action@main
-      - run: echo ${{ steps.coverage.outputs.documented }}
-
   # Swift testing on macOS (Apple Silicon)
   test-swift:
     runs-on: macos-14-xlarge
