+++ src/profile/encrypted/encryption/encryption_scheme_version.rs
@@ -1,0 +1,16 @@
+use crate::prelude::*;
+
+#[repr(u32)]
+#[derive(
+    Serialize_repr,
+    Deserialize_repr,
+    Clone,
+    PartialEq,
+    Eq,
+    Hash,
+    derive_more::Display,
+    derive_more::Debug,
+)]
+pub enum EncryptionSchemeVersion {
+    Version1 = 1,
+}
