+++ src/profile/encrypted/key_derivation/versioned_password_based_key_derivation.rs
@@ -1,0 +1,5 @@
+use crate::prelude::*;
+
+pub trait VersionedPasswordBasedKeyDerivation: VersionOfAlgorithm {
+    fn kdf(&self, password: impl AsRef<str>) -> Exactly32Bytes;
+}
