+++ src/profile/encrypted/mod.rs
@@ -1,11 +1,7 @@
-mod encrypted_profile_snapshot;
-mod encryption_scheme;
-mod password_based_key_derivation_scheme;
+mod encryption;
+mod key_derivation;
 mod versioned_algorithm;
-mod versioned_encryption;
 
-pub use encrypted_profile_snapshot::*;
-pub use encryption_scheme::*;
-pub use password_based_key_derivation_scheme::*;
+pub use encryption::*;
+pub use key_derivation::*;
 pub use versioned_algorithm::*;
-pub use versioned_encryption::*;
